refactor: add `BaseInteractionContext` (#384) and #264
